[fix]: [CCM-12163]: Remove Round off calculation for active spend and licence usage (#46785)

+++ 490-ce-commons/src/main/java/io/harness/ccm/commons/beans/ActiveSpendResultSetDTO.java
@@ -23,7 +23,7 @@ import lombok.experimental.FieldDefaults;
 @EqualsAndHashCode(callSuper = false)
 @OwnedBy(CE)
 public class ActiveSpendResultSetDTO {
-  Long cost;
+  Double cost;
   String month;
   String cloudProvider;
   Long maxDay;
